+++ api/src/main/java/com/goree/api/service/GroupService.java
@@ -23,7 +23,7 @@ public class GroupService {
         Group makedGroup = groupMapper.selectGroupByName(group.getName());
         return makedGroup;
     }
-    public Group findGroupById(int id) {
+    public Group findGroupById(long id) {
         return groupMapper.selectGroupById(id);
     }
     public List<Group> findGroupAll() {
@@ -32,7 +32,7 @@ public class GroupService {
     public Group findGroupByName(String name) {
         return groupMapper.selectGroupByName(name);
     }
-    public void joinMember(int groupId, int memberId) {
+    public void joinMember(long groupId, long memberId) {
         groupMapper.insertToGroupHasMember(groupId, memberId);
     }
 }
